+++ R/find.turn.point.R
@@ -1,7 +1,47 @@
+#' @description
+#' Compute local maxima turn points. 
+#' @param y The y values of a curve in x-y plane.
+#' @param ties.method specifies the method rank uses to break ties.
+#' @return boolean row with local maxima turn point.
+#' @export
+find_local_maxima <- function(y, ties.method) {
+    padded_y <- rev(as.vector(c(-Inf, y, -Inf)))
+
+    # each row is 3 consecutive values in descending order
+    # rows are sorted in ascending order
+    z <- embed(padded_y, dim = 3)
+
+    # reverse the row ordering
+    # first column is equal y
+    z <- z[rev(seq(nrow(z))), ]
+
+    # row where the max is in the middle is a turn point
+    v <- max.col(z, ties.method = ties.method) == 2
+
+    return(v)
+}
+
+#' @description
+#' Compute maxima and minima turn points. 
+#' @param y The y values of a curve in x-y plane.
+#' @return boolean row with local maxima and minima turn points.
+#' @export
+msExtrema <- function(y) {
+    index1 <- find_local_maxima(y, ties.method = "first")
+    index2 <- find_local_maxima(-y, ties.method = "last")
+
+    # this is some sort of safety mechanism to protect against numerical issues
+    index.max <- index1 & !index2
+    index.min <- index2 & !index1
+
+    list(index.max = index.max, index.min = index.min)
+}
+
 #' Find peaks and valleys of a curve.
-#' 
+#'
+#' @description
 #' This is an internal function which finds the peaks and valleys of a smooth curve.
-#' 
+#'
 #' @param y The y values of a curve in x-y plane.
 #' @return A list object:
 #' \itemize{
@@ -11,48 +51,34 @@
 #' @export
 #' @examples
 #' find.turn.point(y)
-find.turn.point <- function(values) {
-  
-  find_extrema <- function(values, ties.method) {
-    # this might not work correctly
-    z <- embed(rev(as.vector(c(-Inf, values,-Inf))), dim = 3)
-    z <- z[rev(seq(nrow(z))),]
-    v <- max.col(z, ties.method = ties.method) == 2
-    return(v)
-  }
-  
-  analyse_extrema <- function(values) {
-    minima <- find_extrema(values, ties.method = "first") # find peaks
-    maxima <- find_extrema(-values, ties.method = "last") # find valleys
-    # remove overlaps
-    maxima.indices <- minima & !maxima
-    minima.indices <- maxima & !minima
-    return(list(minima.indices = minima.indices, maxima.indices = maxima.indices))
-  }
-  
-  values <- values[!is.na(values)]
-  # flat curve => peak is middle and valleys are the first and the last
-  if (length(unique(values)) == 1) {
-    x <- new("list")
-    x$pks <- round(length(values) / 2)
-    x$vlys <- c(1, length(values))
-    return(x)
-  }
-  
-  extrema <- analyse_extrema(values)
-  maxima <- which(extrema$maxima.indices)
-  minima <- which(extrema$minima.indices)
-  # if the first or the last index are not maxima, added them to minima
-  if (maxima[1] != 1)
-    minima <- c(1, minima)
-  if (maxima[length(maxima)] != length(values))
-    minima <- c(minima, length(values))
-  # if there is only one maximum, the minima are the first and the last index
-  # (probably because Gaussian was used)
-  if (length(maxima) == 1)
-    minima <- c(1, length(values))
-  x <- new("list")
-  x$pks <- maxima
-  x$vlys <- minima
-  return(x)
+find.turn.point <- function(y) {
+    y <- y[!is.na(y)] # filter NA values
+
+    if (length(unique(y)) == 1) { # if exactly one distinct value
+        middle_index <- round(length(y) / 2) # get mid index
+        start_and_end <- c(1, length(y)) # get first and last index
+        return(list(pks = middle_index, vlys = start_and_end))
+    } else {
+        b <- msExtrema(y)
+
+        pks <- which(b$index.max)
+        vlys <- which(b$index.min)
+
+        if (pks[1] != 1) {
+            vlys <- c(1, vlys)
+        }
+
+        if (pks[length(pks)] != length(y)) {
+            vlys <- c(vlys, length(y))
+        }
+
+        if (length(pks) == 1) {
+            vlys <- c(1, length(y))
+        }
+
+        x <- new("list")
+        x$pks <- pks
+        x$vlys <- vlys
+        return(x)
+    }
 }