+++ R/prof.to.features.R
@@ -641,8 +641,10 @@ prof.to.features <- function(a,
     if (is.na(max.bw)) max.bw <- diff(range(a[, 2], na.rm = TRUE)) / 15
     if (min.bw >= max.bw) min.bw <- max.bw / 4
 
-    base.curve <- compute_base_curve(a[, 2])
-    all.times <- compute_all_times(base.curve)
+    base.curve <- sort(unique(a[, 2]))
+    base.curve <- cbind(base.curve, base.curve * 0)
+
+    all.times <- compute_delta_rt(base.curve[, 1])
 
     this.features <- matrix(0, nrow = 1, ncol = 5)
     colnames(this.features) <- c("mz", "pos", "sd1", "sd2", "area")